Fix regression introduced by `CacheServer` entry
In a7cdd0a8 (merge request !21 [1]), an entry was added to the
`SINGLE_OPTIONS` array. Due to a missing comma [2], the new entry was
actually not added but merely concatenated to its successor, causing
the `CleanMethod` entry to be changed to `CacheServerCleanMethod`.

That in turn caused e.g. `pycman` to print the following warning:

> /usr/lib/python3.11/site-packages/pycman/config.py:141: InvalidSyntax: unable to parse /etc/pacman.conf, unrecognized option: 'CleanMethod'

Add the missing comma to fix this.

[1]: https://gitlab.archlinux.org/archlinux/pyalpm/-/merge_requests/21
[2]: https://gitlab.archlinux.org/archlinux/pyalpm/-/merge_requests/21/diffs#0c44f0fe1eeb4e111c9fde851ff7eacf8d9df35b_67_67

CC: Jelle van der Waa <jelle@archlinux.org>